Katana Benchmarks (#1467)
* updated dojo benches to work with the new version

* basic bench

* katana bench to separate file

* prefunded account manager

* iterating over prefunded

* working benchmark

* timing responses and cleanup

* allowed for a greater number of prefunded

* spawn before each move

* auth writer fix

* last touches

* formatting

* initial katana runner macro

* working macro

* working katana macro

* changed runner to include accounts

* katana_test test

* deploying contract

* running auth script

* katana runner from a different executable

* working but really slow

* updated contracts in new format

* added a deploy to function like macro

* all benches passing

* updated to new katana

* overall cleanup

* fmt and clippy

* moved other uses of katan runner to macro

* rust fmt fix

* block sizes assert

* block duration from logs

* time for each 2 consecutive transactions

* minor cleanup

* bench for different numbers of tx

* added tps estimation

* benchmark logs now use json

* loading contract address from sozo

* heavy benchmark

* small refactor of benchmarks

* fixed one off account issue

* minor benchmark fixes

* heavy compute benchmarks

* multicall bechmarks

* final

* results aftere rebase

* Readme update and CI fix

* fixed test issues

* deployed using sozo depencency

* moved runner to benches

* parsing json before filtering logs

* changed permalinks to point to the main repo

* moved benchmarks behind a feature

* reverset feature flag for benches

* formatting

* moved runner back to katana

* moved the deployment logic to benches

* skipping benchmarks by default

* formatting after changes

* sync runner init

* fixes after rebase

---------

+++ crates/katana/primitives/src/genesis/allocation.rs
@@ -224,7 +224,7 @@ impl From<DevGenesisAccount> for GenesisAllocation {
 /// A helper type for allocating dev accounts in the genesis block.
 #[must_use]
 pub struct DevAllocationsGenerator {
-    total: u8,
+    total: u16,
     seed: [u8; 32],
     balance: U256,
     class_hash: FieldElement,
@@ -234,7 +234,7 @@ impl DevAllocationsGenerator {
     /// Create a new dev account generator for `total` number of accounts.
     ///
     /// This will return a [DevAllocationsGenerator] with the default parameters.
-    pub fn new(total: u8) -> Self {
+    pub fn new(total: u16) -> Self {
         Self {
             total,
             seed: [0u8; 32],
